feat(nat): index 3 is filled, and the boundary count is 1 of 10, not 3
`Nat.floorRoot`/`Nat.ceilRoot` land as bounded searches -- Mathlib's bodies are
products over `Nat.factorization`, a `Finsupp` this kernel cannot state --
opening `Mathlib.Data.Nat.Factorization.Root` at pool 18 for cycle index 3.
Construction only (ADR-0653): no theorem about either, no fact registered, and
`shape_search --const` is ABSENT for both against a working positive control.

ADR-1220's inherited 3-of-10 boundary count was measured against Mathlib's
definition. Against this one it is 1 of 10: `ceilRoot_one_left` and
`ceilRoot_zero_right` both stop being `Eq.refl`, exactly as ADR-1220 predicted a
search would do. That reading is now a GATED assertion at free variables
(`exactly_one_drawn_row_is_settled_by_reduction`), killed by two mutants, not a
paragraph.

Draw 11's `natural-nth-root` review is REDONE, not amended: the two names take
the recorded `root` sweep from 11 declarations to 13. All 13 read by name. The
finding is that these are DIVISIBILITY roots and `Nat.nthRoot` is an ORDER root
-- measured, since substituting `floorRoot` into `Nat.le_nthRoot_iff` makes it
FALSE at (2, 3, 12). Two awkward things recorded rather than glossed: the stem
sweep is a NAME sweep and would not have fired on a different spelling, and the
divisibility and order adjunction lemmas have rhyming proof skeletons, so the
lane landing the supporting theorems from development must re-screen first.

Eleven mutants, nine killed, TWO SURVIVED and both are reported: `ROOT_HEIGHT`
below `Nat.pow`/`Nat.mod` (admitted, every value unchanged, invisible to every
test -- ADR-1230's third outcome in definition shape), and an extensionally
identical fuel bound, which should survive. M10's run did not complete and its
other three results are unknown rather than passing. M5 was predicted to survive
and did not; the source comment records the correction.

Verification: `--lib nat_prelude::factorization_root` 6 passed / 0 failed;
`--lib nat_prelude::` 309 passed / 0 failed, up from 303; clippy clean.
Environment 2706 -> 2708, exactly +2 (`definition 372 -> 374`), `axiom=30`
unchanged. Manifest byte-identical under the refreshed snapshot, 420 entries,
`development=160 held-out=150 train=110`. Five gates green before and after,
`held_out=166 settled=0 PASS` both times; no fact moved partition.

Agent: unblock-index-three

+++ crates/axeyum-lean-kernel/src/nat_prelude/factorization_root.rs
@@ -124,7 +124,7 @@ use crate::env::ReducibilityHint;
 
 /// Delta height for both definitions. Strictly above `Nat.pow` (3),
 /// `Nat.mod` (3) and `Nat.beq`, the only `Nat` definitions either one calls.
-const ROOT_HEIGHT: u32 = 6;
+const ROOT_HEIGHT: u16 = 6;
 
 /// Declare `Nat.floorRoot` and `Nat.ceilRoot`. Definitions only — see this
 /// module's doc for why no theorem about either is declared here.
